pyproject.toml
[project]
name = "docling-agent-assistant"
version = "2.0.0"
description = "Catalogue de prix BTP — extraction IA de factures"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0",
    "pandas>=2.0",
    "streamlit==1.38.0",
    "openpyxl>=3.1",
    "pillow>=10.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP"]
ignore = ["E501"]
